refactor names & project structure
- empowerment function in __init__.py for a more sensible import statement
- empowerment -> compute_empowerment
- info_theory -> information_theory
- update README
