+++ example_use.py
@@ -1,6 +1,6 @@
-import mazeworld
-from mazeworld import MazeWorld
-from empowerment import empowerment, EmpMaxAgent
+from empowerment import mazeworld
+from empowerment.mazeworld import MazeWorld
+from empowerment import compute_empowerment, EmpowermentMaximiser 
 import numpy as np 
 import matplotlib.pyplot as plt
 import time 
@@ -26,21 +26,21 @@ def example_1():
     maze.add_wall( (8, 4), "W")
     maze.add_wall( (8, 3), "N")
     # compute the 5-step empowerment at each cell 
-    E = maze.empowerment(n_step=5)
+    E = maze.compute_empowerment(n_step=5)
     # plot the maze world
     maze.plot(colorMap=E)
     plt.title('5-step empowerment')
     plt.show()
 
 def example_2():
-    """ builds grid world with doors and plots empowerment landscape """ 
+    """ Builds grid world with doors and plots empowerment landscape """ 
     maze = MazeWorld(8,8)
     for i in range(maze.width):
         if i is not 6 : maze.add_wall([2, i], "N") 
     for i in range(maze.width):
         if i is not 2 : maze.add_wall([5, i], "N")
     n_step = 4
-    E = maze.empowerment(n_step=n_step, n_samples=8000)
+    E = maze.compute_empowerment(n_step=n_step, n_samples=8000)
     maze.plot(colorMap=E)
     plt.title('%i-step empowerment' % n_step)
     plt.show()
@@ -59,9 +59,9 @@ def example_3():
     s =  maze._cell_to_index(initpos)
     T = emptymaze.compute_model()
     B = maze.compute_model()
-    E = maze.empowerment(n_step = n_step).reshape(-1)
+    E = maze.compute_empowerment(n_step = n_step).reshape(-1)
     n_s, n_a, _ = T.shape
-    agent = EmpMaxAgent(alpha=0.1, gamma=0.9, T = T, n_step=n_step, n_samples=1000, det=1.)
+    agent = EmpowermentMaximiser(alpha=0.1, gamma=0.9, T = T, n_step=n_step, n_samples=1000, det=1.)
     steps = int(10000) 
     visited = np.zeros(maze.dims)
     tau = np.zeros(steps)
