+++ setup.py
@@ -1,9 +1,9 @@
 from setuptools import setup
  
 setup(
-    name='empowerment',       # This is the name of your PyPI-package.
-    version='0.1',          # Update the version number for new releases
-    description='Compute the information-theoretic empowerment in toy worlds',
+    name='empowerment',
+    version='0.1',
+    description='Compute the information-theoretic empowerment in grid worlds',
     url='http://github.com/Mchristos/empowerment',
     author='Chris',
     license='MIT',
